requirements.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.8
Django==4.2.9
djangorestframework==3.14.0
exceptiongroup==1.2.0
filelock==3.13.1
identify==2.5.33
iniconfig==2.0.0
nodeenv==1.8.0
packaging==23.2
platformdirs==4.1.0
pluggy==1.4.0
pre-commit==3.6.0
pytest==8.0.0
pytest-django==4.8.0
pytz==2023.4
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
virtualenv==20.25.0
